requirements.txt
langchain
#openai==0.28.1
openai==1.12.0
streamlit
python-dotenv
tiktoken
cohere
faiss-cpu
streamlit_chat
llmonitor
pinecone-client
pypdf
lida
torch
lida[transformers]
streamlit_feedback
langfuse
